Relocating style constants to lib directory
