+++ src/app/components/Footer/index.jsx
@@ -1,6 +1,11 @@
 import React from 'react';
 import styled from 'styled-components';
-import { C_STORM, C_WHITE, FF_NEWS, GEL_SPACING_DBL } from '../../constants';
+import {
+  C_STORM,
+  C_WHITE,
+  FF_NEWS,
+  GEL_SPACING_DBL,
+} from '../../../lib/constants/styles';
 
 const StyledFooter = styled.footer`
   background-color: ${C_STORM};
